+++ src/types/deasync.d.ts
@@ -1,0 +1,4 @@
+declare module "deasync" {
+    function deasync<T>(promise: Promise<T>): T
+    export = deasync
+}